Fix bugs and resolve compiler warnings from code review
Critical fixes:
- replay_buffer: Fix race condition in ReplayBuffer_Shutdown (audio cleanup inside CS)
- replay_buffer: Reset g_audioMaxDuration between sessions
- nvenc_encoder: Remove sync mode fallback - async is now required

Memory/resource fixes:
- capture: Store dwmapi.dll handle in static to prevent leak
- encoder: Remove debug fopen() calls that bypassed --debug flag

Warning fixes:
- aac_encoder: Change configSize to UINT32 to match MF API (C4057)
- overlay: Explicit BYTE cast for overlayAlpha (C4244)
- overlay: Remove unused pGdipGetImageWidth/Height (C4189)
- overlay: Rename shadowed bgColor to btnBgColor (C4456)
- overlay: Rename shadowed mins/secs to calcMins/calcSecs (C4456)
- overlay: Remove unused audioDivider variable (C4189)
- replay_buffer: Add (void)userData to suppress C4100
- replay_buffer: Rename shadowed realElapsedSec to logElapsedSec (C4456)
- sample_buffer: Fix eviction counter pattern
- audio_capture: Remove unnecessary peakInitialized flag

+++ src/aac_encoder.c
@@ -39,7 +39,7 @@ struct AACEncoder {
     
     // Encoder config (AudioSpecificConfig)
     BYTE* configData;
-    int configSize;
+    UINT32 configSize;  // UINT32 to match Media Foundation API
     
     BOOL initialized;
 };
@@ -383,7 +383,7 @@ BOOL AACEncoder_GetConfig(AACEncoder* encoder, BYTE** configData, int* configSiz
     
     if (encoder->configData && encoder->configSize > 0) {
         *configData = encoder->configData;
-        *configSize = encoder->configSize;
+        *configSize = (int)encoder->configSize;  // Safe cast - config is small
         return TRUE;
     }
